requirements.txt
Markdown==3.3.7
mkdocs-material==8.4.2
mkdocs-with-pdf==0.9.3
pyyaml==6.0
click==8.1.3
jinja2==3.1.2
